+++ dzVents/runtime/HTTPResponse.lua
@@ -8,23 +8,15 @@ local function HTTPResponce(domoticz, responseData)
 	self.headers = responseData.headers or {}
 	self.baseType = domoticz.BASETYPE_HTTP_RESPONSE
 	self.data = responseData.data
-	self.status = responseData.status
 	self._contentType = _.get(self.headers, {'Content-Type'}, '')
+    
+    self.isJSON = false
 
-	self.isJSON = false
-
-	if self.status and self.status ~= ""  then
-		local chunks = utils.stringSplit(self.status)
-		self.protocol = chunks[1]
-		self.statusCode = tonumber(chunks[2] )
-		self.statusMessage = table.concat(_.slice(chunks,3)," ")
-	else
-		self.statusMessage = ""
-		self.protocol = ""
-		self.statusCode = 0
-	end
-
-	self.ok = ( self.statusCode >= 200 and self.statusCode <= 299 )
+    self.statusText = responseData.statusText
+    self.protocol = responseData.protocol
+    self.statusCode = responseData.statusCode
+    
+    self.ok = ( self.statusCode >= 200 and self.statusCode <= 299 )
 	
 	self.isHTTPResponse = true
 	self.isDevice = false